libsel4arm-vmm: dev_paddr_in_range uses uintptr_t
Updated 'dev_paddr_in_range' to use a uintptr_t addresses rather
than a 32bit address. This allows for creating devices in the
64bit range.

+++ libsel4arm-vmm/src/devices.h
@@ -88,7 +88,7 @@ void *map_emulated_device(vm_t *vm, const struct device *d);
  * @param[in] d    A description of the device to check
  * @return         non-zero if the address belongs to the device
  */
-static inline int dev_paddr_in_range(uint32_t addr, const struct device *d)
+static inline int dev_paddr_in_range(uintptr_t addr, const struct device *d)
 {
     return ((addr >= d->pstart) && addr < (d->pstart + d->size));
 }
